requirements.txt
aiogram==3.27.0
aiosqlite==0.22.1
aiohttp==3.13.5
python-dotenv==1.2.2
Pillow==12.2.0
